+++ lightning.info.yml
@@ -43,6 +43,7 @@ dependencies:
   - lightning_workflow
   - lightning_contact_form
   - lightning_page
+  - lightning_roles
   - lightning_search
   - lightning_landing_page
   - lightning_media_document
