+++ modules/lightning_features/lightning_core/lightning_core.info.yml
@@ -14,4 +14,5 @@ dependencies:
 components:
   - lightning_contact_form
   - lightning_page
+  - lightning_roles
   - lightning_search
