+++ modules/lightning_features/lightning_core/lightning_core.services.yml
@@ -1,10 +1,4 @@
 services:
-  lightning.content_roles:
-    class: '\Drupal\lightning_core\ContentRoleManager'
-    arguments:
-      - '@config.factory'
-      - '@entity_type.manager'
-
   lightning.route_subscriber:
     class: '\Drupal\lightning_core\Routing\RouteSubscriber'
     arguments:
