+++ modules/lightning_features/lightning_core/modules/lightning_roles/lightning_roles.info.yml
@@ -1,0 +1,6 @@
+name: 'Lightning Roles'
+core: 8.x
+type: module
+package: Lightning
+description: 'Automatically manages user roles and permissions.'
+version: '8.x-2.05'
