+++ modules/lightning_features/lightning_core/modules/lightning_roles/lightning_roles.services.yml
@@ -1,0 +1,6 @@
+services:
+  lightning.content_roles:
+    class: '\Drupal\lightning_roles\ContentRoleManager'
+    arguments:
+      - '@config.factory'
+      - '@entity.query'
