+++ modules/lightning_features/lightning_layout/config/optional/user.role.layout_manager.yml
@@ -6,3 +6,6 @@ is_admin: false
 permissions:
   - administer node display
   - administer panelizer
+dependencies:
+  module:
+    - lightning_roles
