+++ modules/lightning_features/lightning_media/config/optional/user.role.media_creator.yml
@@ -1,6 +1,8 @@
 langcode: en
 status: true
-dependencies: {  }
+dependencies:
+  module:
+    - lightning_roles
 id: media_creator
 label: 'Media creator'
 weight: 5
