+++ modules/lightning_features/lightning_media/config/optional/user.role.media_manager.yml
@@ -1,6 +1,8 @@
 langcode: en
 status: true
-dependencies: {  }
+dependencies:
+  module:
+    - lightning_roles
 id: media_manager
 label: 'Media manager'
 weight: 6
