+++ modules/lightning_features/lightning_preview/config/optional/user.role.workspace_reviewer.yml
@@ -1,6 +1,8 @@
 langcode: en
 status: true
-dependencies: {  }
+dependencies:
+  module:
+    - lightning_roles
 id: workspace_reviewer
 label: 'Workspace reviewer'
 weight: 11
